fix memcached integration

+++ src/DDTrace/Integrations/Memcached/MemcachedIntegration.php
@@ -60,64 +60,111 @@ class MemcachedIntegration extends Integration
 
         // bool Memcached::add ( string $key , mixed $value [, int $expiration ] )
         dd_trace('Memcached', 'add', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'add', func_get_args());
         });
 
         // bool Memcached::addByKey ( string $server_key , string $key , mixed $value [, int $expiration ] )
         dd_trace('Memcached', 'addByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'addByKey', func_get_args());
         });
 
         // bool Memcached::append ( string $key , string $value )
         dd_trace('Memcached', 'append', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'append', func_get_args());
         });
 
         // bool Memcached::appendByKey ( string $server_key , string $key , string $value )
         dd_trace('Memcached', 'appendByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'appendByKey', func_get_args());
         });
 
         // bool Memcached::cas ( float $cas_token , string $key , mixed $value [, int $expiration ] )
         dd_trace('Memcached', 'cas', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCas($this, func_get_args());
         });
 
         // bool Memcached::casByKey ( float $cas_token , string $server_key , string $key , mixed $value
         //     [, int $expiration ] )
         dd_trace('Memcached', 'casByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCasByKey($this, func_get_args());
         });
 
         // int Memcached::decrement ( string $key [, int $offset = 1 [, int $initial_value = 0
         //     [, int $expiry = 0 ]]] )
         dd_trace('Memcached', 'decrement', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'decrement', func_get_args());
         });
 
         // int Memcached::decrementByKey ( string $server_key , string $key
         //      [, int $offset = 1 [, int $initial_value = 0 [, int $expiry = 0 ]]] )
         dd_trace('Memcached', 'decrementByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'decrementByKey', func_get_args());
         });
 
         // bool Memcached::delete ( string $key [, int $time = 0 ] )
         dd_trace('Memcached', 'delete', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'delete', func_get_args());
         });
 
         // bool Memcached::deleteByKey ( string $server_key , string $key [, int $time = 0 ] )
         dd_trace('Memcached', 'deleteByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'deleteByKey', func_get_args());
         });
 
         // array Memcached::deleteMulti ( array $keys [, int $time = 0 ] )
         dd_trace('Memcached', 'deleteMulti', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceMulti($this, 'deleteMulti', func_get_args());
         });
 
         // bool Memcached::deleteMultiByKey ( string $server_key , array $keys [, int $time = 0 ] )
         dd_trace('Memcached', 'deleteMultiByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
             return MemcachedIntegration::traceMultiByKey($this, 'deleteMultiByKey', func_get_args());
         });
 
@@ -228,10 +275,6 @@ class MemcachedIntegration extends Integration
     public static function traceCommand($memcached, $command, $args)
     {
         $tracer = GlobalTracer::get();
-        if ($tracer->limited()) {
-            return dd_trace_forward_call();
-        }
-
         $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             "Memcached.$command"
@@ -255,10 +298,6 @@ class MemcachedIntegration extends Integration
     public static function traceCommandByKey($memcached, $command, $args)
     {
         $tracer = GlobalTracer::get();
-        if ($tracer->limited()) {
-            return dd_trace_forward_call();
-        }
-
         $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             "Memcached.$command"
@@ -281,10 +320,6 @@ class MemcachedIntegration extends Integration
     public static function traceCas($memcached, $args)
     {
         $tracer = GlobalTracer::get();
-        if ($tracer->limited()) {
-            return dd_trace_forward_call();
-        }
-
         $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             'Memcached.cas'
@@ -305,10 +340,6 @@ class MemcachedIntegration extends Integration
     public static function traceCasByKey($memcached, $args)
     {
         $tracer = GlobalTracer::get();
-        if ($tracer->limited()) {
-            return dd_trace_forward_call();
-        }
-
         $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             'Memcached.casByKey'
@@ -331,10 +362,6 @@ class MemcachedIntegration extends Integration
     public static function traceMulti($memcached, $command, $args)
     {
         $tracer = GlobalTracer::get();
-        if ($tracer->limited()) {
-            return dd_trace_forward_call();
-        }
-
         $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             "Memcached.$command"
@@ -354,10 +381,6 @@ class MemcachedIntegration extends Integration
     public static function traceMultiByKey($memcached, $command, $args)
     {
         $tracer = GlobalTracer::get();
-        if ($tracer->limited()) {
-            return dd_trace_forward_call();
-        }
-
         $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             "Memcached.$command"
